refactor/clint: invoiceOverview + new component InvoiceRow.tsx
